# backend/api/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a.",
        "title": "Manager API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "Manager API Support",
            "url": "https://github.com/Inteli-College/2024-1B-T02-EC10-G04",
            "email": "gomedicine@inteli.edu.br"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/pyxis": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieve all Pyxis entities",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Pyxis"
                ],
                "summary": "Retrieve all Pyxis entities",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/dto.FindPyxisOutputDTO"
                            }
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Create a new Pyxis entity",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Pyxis"
                ],
                "summary": "Create a new Pyxis entity",
                "parameters": [
                    {
                        "description": "Pyxis entity to create",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CreatePyxisInputDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.CreatePyxisOutputDTO"
                        }
                    }
                }
            }
        },
        "/pyxis/qrcode": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Create a QR code for a given pyxis ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "image/png"
                ],
                "tags": [
                    "Pyxis"
                ],
                "summary": "Generate a QR code for a Pyxis",
                "parameters": [
                    {
                        "description": "Pyxis ID to generate QR code for",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.GenerateQRCodeOutputDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "QR code image",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/pyxis/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieve a Pyxis entity by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Pyxis"
                ],
                "summary": "Retrieve a Pyxis entity by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Pyxis ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.FindPyxisOutputDTO"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Update a Pyxis entity",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Pyxis"
                ],
                "summary": "Update a Pyxis entity",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Pyxis ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Pyxis entity to update",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.UpdatePyxisInputDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.UpdatePyxisOutputDTO"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Delete a Pyxis entity",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Pyxis"
                ],
                "summary": "Delete a Pyxis entity",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Pyxis ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/pyxis/{id}/medicines": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get all medicines related to a Pyxis",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Pyxis"
                ],
                "summary": "Get medicines from a Pyxis",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Pyxis ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/dto.FindMedicineOutputDTO"
                            }
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Disassociate a sequence n of medicines from a Pyxis",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Pyxis"
                ],
                "summary": "Disassociate medicines from a Pyxis",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Pyxis ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Medicines to disassociate",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.DisassociateMedicineInputDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/pyxis/{id}/register-medicine": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Register a existing medicine to a existing Pyxis entity",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Pyxis"
                ],
                "summary": "Register a to a Pyxis entity",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Pyxis ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Medicines to register into Pyxis",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.RegisterMedicinePyxisInputDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "message\": message}",
                        "schema": {
                            "type": "objetct"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dto.CreatePyxisInputDTO": {
            "type": "object",
            "properties": {
                "label": {
                    "type": "string"
                }
            }
        },
        "dto.CreatePyxisOutputDTO": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "label": {
                    "type": "string"
                }
            }
        },
...
        "dto.DisassociateMedicineInputDTO": {
            "type": "object",
            "properties": {
                "medicines": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "dto.FindMedicineOutputDTO": {
            "type": "object",
            "properties": {
                "batch": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "stripe": {
                    "$ref": "#/definitions/entity.StripeType"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "dto.FindPyxisOutputDTO": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "label": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "dto.GenerateQRCodeOutputDTO": {
            "type": "object",
            "properties": {
                "pyxis_id": {
                    "type": "string"
                }
            }
        },
...
        "dto.RegisterMedicinePyxisInputDTO": {
            "type": "object",
            "properties": {
                "medicines": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
...
        "dto.UpdatePyxisInputDTO": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "label": {
                    "type": "string"
                }
            }
        },
        "dto.UpdatePyxisOutputDTO": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "label": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "entity.StripeType": {
            "type": "string",
            "enum": [
                "red",
                "yellow",
                "black"
            ],
            "x-enum-varnames": [
                "StripeRed",
                "StripeYellow",
                "StripeBlack"
            ]
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "\"Type: Bearer token\"",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}